+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] 2025-07-19
+
+### Changed
+
+- Update cmake and make build and conan integration
+- Remove CMakeUserPresets
+- Add libtorch 2.7.1 dependency
+- Fix conan package version
+- Add conan integration for dependency management
+- Remove git submodules
+
 ## [1.1.1] 2024-12-13
 
 - Added a new parameter `quiet` to enable/disable the warning messages in the Stratified K-Fold partitioning. Default value `true`.